requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
argh==0.26.2
async-timeout==3.0.1
attrs==20.3.0
black==20.8b1
cffi==1.14.4
chardet==3.0.4
click==7.1.2
discord.py==1.5.1
flake8==3.8.4
idna==2.10
mccabe==0.6.1
multidict==4.7.6
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyNaCl==1.4.0
python-dotenv==0.15.0
PyYAML==5.3.1
regex==2020.11.13
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
watchgod==0.6
yarl==1.5.1
